fix: PREV-72 - Thumbnail generation goes in timeout (#13)
Thumbnail generation with specific PDFs caused the request to the preview service to go in timeout. This was caused by LibreOffice taking minutes to convert multi layered or simply resource heavy PDFs. This problem was tested and reproduced even calling the LibreOffice command line utilities.

To fix this problem it was decided to relived LibreOffice from some load and use another dependency to handle PDFs -> Image conversion. PDFium was chosen, more specifically pypdfium2. This library is also faster and more resource efficient.

It was also added a template model for the next PR to follow strictly